Implementa roles de usuário com enum e permissões iniciais para atendente

+++ app/controllers/reservas_controller.rb
@@ -1,21 +1,19 @@
 class ReservasController < ApplicationController
 
     layout "layout_reserva" # Define um layout exclusivo
-
-
     before_action :set_reserva, only: [:show, :edit, :update, :destroy]
   before_action :authenticate_user!
-
 
   # Lista todas as reservas
   def index
     @reservas = current_user.reservas.order(data: :asc) if current_user
   end
-  
+
 
   # Exibe uma reserva específica
   def show
   end
+
 
   # Exibe o formulário de nova reserva
   def new
@@ -27,13 +25,13 @@ class ReservasController < ApplicationController
   def create
     @reserva = Reserva.new(reserva_params)
     @reserva.user = current_user  # A reserva pertence ao usuário logado
-
     if @reserva.save
       redirect_to @reserva, notice: 'Reserva criada com sucesso!'
     else
       render :new, status: :unprocessable_entity
     end
   end
+
 
   # Exibe o formulário para editar uma reserva
   def edit
@@ -50,27 +48,29 @@ class ReservasController < ApplicationController
   end
 
    # Exclui uma reserva
-   def destroy
+def destroy
+  # Verifica se o usuário logado é um atendente ou o dono da reserva
+  if current_user.atendente? || @reserva.user == current_user
     if @reserva.destroy
       redirect_to reservas_url, notice: 'Reserva excluída com sucesso!'
     else
       redirect_to reservas_url, alert: 'Erro ao excluir a reserva.'
     end
+  else
+    redirect_to reservas_url, alert: 'Você não tem permissão para excluir esta reserva.'
   end
-  
+end
 
 
-  private
 
   # Busca a reserva pelo ID
   def set_reserva
     @reserva = current_user.reservas.find(params[:id])
   end
-  
+
 
   # Define os parâmetros permitidos
   def reserva_params
     params.require(:reserva).permit(:data, :status)
   end
-
 end
